requirements.txt
apipkg==1.4
autopep8==1.2.4
cycler==0.10.0
dbfread==2.0.6
execnet==1.4.1
googlemaps==2.4.4
matplotlib==1.5.3
numpy==1.11.2
pandas==0.19.0
pep8==1.7.0
psycopg2==2.6.2
py==1.4.31
pyparsing==2.1.9
pytest==3.0.3
pytest-cache==1.0
pytest-pep8==1.0.6
python-dateutil==2.5.3
pytz==2016.7
PyYAML==3.12
requests==2.10.0
scikit-learn==0.18
scipy==0.18.1
seaborn==0.7.1
six==1.10.0
sklearn==0.0
SQLAlchemy==1.0.15
xlrd==1.0.0
